2024-02-11 009: Move CompilationEnclosure
