+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/comp/internal/CompilationImpl.java
@@ -23,7 +23,7 @@ import tripleo.elijah.comp.i.extra.CompilerInputListener;
 import tripleo.elijah.comp.i.extra.IPipelineAccess;
 import tripleo.elijah.comp.impl.DefaultCompilationEnclosure;
 import tripleo.elijah.comp.impl.LCM_Event_RootCI;
-import tripleo.elijah.comp.internal_move_soon.CompilationEnclosure;
+import tripleo.elijah_elevated.comp.backbone.CompilationEnclosure;
 import tripleo.elijah.comp.nextgen.CP_Paths__;
 import tripleo.elijah.comp.nextgen.i.CPX_CalculateFinishParse;
 import tripleo.elijah.comp.nextgen.i.CP_Paths;
