+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/comp/internal/DefaultCompilationBus.java
@@ -1,20 +1,36 @@
 package tripleo.elijah.comp.internal;
 
 import lombok.Getter;
-
-import org.awaitility.core.ConditionTimeoutException;
 import org.jetbrains.annotations.NotNull;
+
 import tripleo.elijah.comp.Compilation;
-import tripleo.elijah.comp.i.*;
-import tripleo.elijah.comp.internal_move_soon.CompilationEnclosure;
+import tripleo.elijah.comp.i.CB_Action;
+import tripleo.elijah.comp.i.CB_Monitor;
+import tripleo.elijah.comp.i.CB_Process;
+import tripleo.elijah.comp.i.CompFactory;
+import tripleo.elijah.comp.i.CompilationChange;
+import tripleo.elijah.comp.i.CompilerDriver;
+import tripleo.elijah.comp.i.ICompilationBus;
+import tripleo.elijah.comp.i.ILazyCompilerInstructions;
+import tripleo.elijah.comp.i.IProgressSink;
+import tripleo.elijah.comp.i.ProgressSinkComponent;
+
 import tripleo.elijah.util.SimplePrintLoggerToRemoveSoon;
 
-import java.lang.reflect.InvocationTargetException;
-import java.util.*;
-import java.util.concurrent.*;
+import tripleo.elijah_elevated.comp.backbone.CompilationEnclosure;
+import tripleo.elijah_elevated.comp.compilation_bus.SingleActionProcess;
 
+import java.lang.reflect.InvocationTargetException;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import java.util.Queue;
+import java.util.concurrent.ConcurrentLinkedQueue;
+
+import java.util.concurrent.TimeUnit;
+import org.awaitility.core.ConditionTimeoutException;
 import static org.awaitility.Awaitility.await;
-import static tripleo.elijah.util.Helpers.List_of;
 
 public class DefaultCompilationBus implements ICompilationBus {
 	private final CB_Monitor _monitor;
