+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/comp/internal/PW_CompilerController.java
@@ -7,7 +7,7 @@ import java.util.concurrent.Flow.Subscriber;
 import io.smallrye.mutiny.Multi;
 import tripleo.elijah.Eventual;
 import tripleo.elijah.comp.i.*;
-import tripleo.elijah.comp.internal_move_soon.CompilationEnclosure;
+import tripleo.elijah_elevated.comp.backbone.CompilationEnclosure;
 import tripleo.elijah.comp.nextgen.i.CP_Paths;
 import tripleo.elijah.comp.nextgen.pw.PW_Controller;
 import tripleo.elijah.comp.nextgen.pw.PW_PushWork;
