+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/nextgen/inputtree/EIT_ModuleInputImpl.java
@@ -5,8 +5,7 @@ import org.jetbrains.annotations.NotNull;
 import tripleo.elijah.ci.*;
 import tripleo.elijah.comp.Compilation;
 import tripleo.elijah.comp.EvaPipeline;
-import tripleo.elijah.comp.internal.*;
-import tripleo.elijah.comp.internal_move_soon.CompilationEnclosure;
+import tripleo.elijah_elevated.comp.backbone.CompilationEnclosure;
 import tripleo.elijah.comp.nextgen.inputtree.EIT_ModuleInput;
 import tripleo.elijah.comp.notation.GM_GenerateModule;
 import tripleo.elijah.comp.notation.GM_GenerateModuleRequest;
