+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/stages/deduce/DeduceTypes2.java
@@ -20,7 +20,6 @@ import org.jetbrains.annotations.*;
 import tripleo.elijah.*;
 import tripleo.elijah.comp.i.*;
 import tripleo.elijah.comp.i.extra.*;
-import tripleo.elijah.comp.internal_move_soon.*;
 import tripleo.elijah.diagnostic.*;
 import tripleo.elijah.g.*;
 import tripleo.elijah.lang.i.*;
@@ -44,6 +43,7 @@ import tripleo.elijah.stages.inter.*;
 import tripleo.elijah.stages.logging.*;
 import tripleo.elijah.util.*;
 import tripleo.elijah.work.*;
+import tripleo.elijah_elevated.comp.backbone.CompilationEnclosure;
 
 import java.util.*;
 import java.util.function.*;
