+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/stages/deduce/pipeline_impl/DeducePipelineImpl.java
@@ -2,7 +2,7 @@ package tripleo.elijah.stages.deduce.pipeline_impl;
 
 import org.jetbrains.annotations.*;
 import tripleo.elijah.comp.*;
-import tripleo.elijah.comp.internal_move_soon.CompilationEnclosure;
+import tripleo.elijah_elevated.comp.backbone.CompilationEnclosure;
 import tripleo.elijah.comp.i.extra.IPipelineAccess;
 
 import java.util.*;
