+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/world/impl/DefaultWorldModule.java
@@ -2,7 +2,7 @@ package tripleo.elijah.world.impl;
 
 import org.jetbrains.annotations.NotNull;
 import tripleo.elijah.*;
-import tripleo.elijah.comp.internal_move_soon.CompilationEnclosure;
+import tripleo.elijah_elevated.comp.backbone.CompilationEnclosure;
 import tripleo.elijah.comp.nextgen.inputtree.EIT_ModuleInput;
 import tripleo.elijah.comp.notation.GN_PL_Run2;
 import tripleo.elijah.lang.i.OS_Module;
