+++ elevated-potential.core.impl/src/main/java/tripleo/elijah_elevated/comp/backbone/CompilationEnclosure.java
@@ -1,4 +1,4 @@
-package tripleo.elijah.comp.internal_move_soon;
+package tripleo.elijah_elevated.comp.backbone;
 
 import io.reactivex.rxjava3.annotations.*;
 import org.jetbrains.annotations.*;
