+++ elevated-potential.core.impl/src/test/java/tripleo/elijah/stages/deduce/DeduceTypesTest.java
@@ -15,7 +15,6 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import tripleo.elijah.comp.*;
 import tripleo.elijah.comp.internal.*;
-import tripleo.elijah.comp.internal_move_soon.*;
 import tripleo.elijah.contexts.*;
 import tripleo.elijah.g.*;
 import tripleo.elijah.lang.i.*;
@@ -33,6 +32,7 @@ import tripleo.elijah.stages.gen_fn.IdentTableEntry;
 import tripleo.elijah.test_help.Boilerplate;
 import tripleo.elijah.world.i.WorldModule;
 import tripleo.elijah.world.impl.DefaultWorldModule;
+import tripleo.elijah_elevated.comp.backbone.CompilationEnclosure;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
